safari-534.53-branch/LayoutTests/http/tests/security/no-popup-from-sandbox.html

<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.setCanOpenWindows(true);
}
</script>
<p>To run this test outside of DumpRenderTree, please disable your popup blocker!</p>
<iframe sandbox="allow-scripts"
  src="data:text/html,
       <script>
       var win = window.open('about:blank', '_blank');
       alert(win ? 'FAIL' : 'PASS');
       </script>"
  ></iframe>
