LayoutTests/http/tests/security/no-popup-from-sandbox-expected.txt

CONSOLE MESSAGE: Blocked opening 'about:blank' in a new window because the request was made in a sandboxed frame whose 'allow-popups' permission is not set.
ALERT: PASS
To run this test outside of DumpRenderTree, please disable your popup blocker!
